Extract OrcTypeConverter utility class
