+++ plugin/trino-iceberg/src/main/java/io/trino/plugin/iceberg/TypeConverter.java
@@ -14,11 +14,6 @@
 package io.trino.plugin.iceberg;
 
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMap;
-import io.trino.orc.metadata.ColumnMetadata;
-import io.trino.orc.metadata.OrcColumnId;
-import io.trino.orc.metadata.OrcType;
-import io.trino.orc.metadata.OrcType.OrcTypeKind;
 import io.trino.spi.TrinoException;
 import io.trino.spi.type.ArrayType;
 import io.trino.spi.type.BigintType;
@@ -41,12 +36,10 @@ import io.trino.spi.type.TypeSignatureParameter;
 import io.trino.spi.type.UuidType;
 import io.trino.spi.type.VarbinaryType;
 import io.trino.spi.type.VarcharType;
-import org.apache.iceberg.Schema;
 import org.apache.iceberg.types.Types;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.atomic.AtomicInteger;
 
@@ -61,11 +54,6 @@ import static java.lang.String.format;
 
 public final class TypeConverter
 {
-    public static final String ORC_ICEBERG_ID_KEY = "iceberg.id";
-    public static final String ORC_ICEBERG_REQUIRED_KEY = "iceberg.required";
-    public static final String ICEBERG_LONG_TYPE = "iceberg.long-type";
-    public static final String ICEBERG_BINARY_TYPE = "iceberg.binary-type";
-
     private TypeConverter() {}
 
     public static Type toTrinoType(org.apache.iceberg.types.Type type, TypeManager typeManager)
@@ -245,143 +233,5 @@ public final class TypeConverter
             return;
         }
         throw new IllegalArgumentException("Either a column identity or nextFieldId is expected");
-    }
-
-    public static ColumnMetadata<OrcType> toOrcType(Schema schema)
-    {
-        return new ColumnMetadata<>(toOrcStructType(0, schema.asStruct(), ImmutableMap.of()));
-    }
-
-    private static List<OrcType> toOrcType(int nextFieldTypeIndex, org.apache.iceberg.types.Type type, Map<String, String> attributes)
-    {
-        switch (type.typeId()) {
-            case BOOLEAN:
-                return ImmutableList.of(new OrcType(OrcTypeKind.BOOLEAN, ImmutableList.of(), ImmutableList.of(), Optional.empty(), Optional.empty(), Optional.empty(), attributes));
-            case INTEGER:
-                return ImmutableList.of(new OrcType(OrcTypeKind.INT, ImmutableList.of(), ImmutableList.of(), Optional.empty(), Optional.empty(), Optional.empty(), attributes));
-            case LONG:
-                return ImmutableList.of(new OrcType(OrcTypeKind.LONG, ImmutableList.of(), ImmutableList.of(), Optional.empty(), Optional.empty(), Optional.empty(), attributes));
-            case FLOAT:
-                return ImmutableList.of(new OrcType(OrcTypeKind.FLOAT, ImmutableList.of(), ImmutableList.of(), Optional.empty(), Optional.empty(), Optional.empty(), attributes));
-            case DOUBLE:
-                return ImmutableList.of(new OrcType(OrcTypeKind.DOUBLE, ImmutableList.of(), ImmutableList.of(), Optional.empty(), Optional.empty(), Optional.empty(), attributes));
-            case DATE:
-                return ImmutableList.of(new OrcType(OrcTypeKind.DATE, ImmutableList.of(), ImmutableList.of(), Optional.empty(), Optional.empty(), Optional.empty(), attributes));
-            case TIME:
-                attributes = ImmutableMap.<String, String>builder()
-                        .putAll(attributes)
-                        .put(ICEBERG_LONG_TYPE, "TIME")
-                        .buildOrThrow();
-                return ImmutableList.of(new OrcType(OrcTypeKind.LONG, ImmutableList.of(), ImmutableList.of(), Optional.empty(), Optional.empty(), Optional.empty(), attributes));
-            case TIMESTAMP:
-                OrcTypeKind timestampKind = ((Types.TimestampType) type).shouldAdjustToUTC() ? OrcTypeKind.TIMESTAMP_INSTANT : OrcTypeKind.TIMESTAMP;
-                return ImmutableList.of(new OrcType(timestampKind, ImmutableList.of(), ImmutableList.of(), Optional.empty(), Optional.empty(), Optional.empty(), attributes));
-            case STRING:
-                return ImmutableList.of(new OrcType(OrcTypeKind.STRING, ImmutableList.of(), ImmutableList.of(), Optional.empty(), Optional.empty(), Optional.empty(), attributes));
-            case FIXED:
-                return ImmutableList.of(new OrcType(OrcTypeKind.BINARY, ImmutableList.of(), ImmutableList.of(), Optional.empty(), Optional.empty(), Optional.empty(), attributes));
-            case BINARY:
-                return ImmutableList.of(new OrcType(OrcTypeKind.BINARY, ImmutableList.of(), ImmutableList.of(), Optional.empty(), Optional.empty(), Optional.empty(), attributes));
-            case DECIMAL:
-                Types.DecimalType decimalType = (Types.DecimalType) type;
-                return ImmutableList.of(new OrcType(OrcTypeKind.DECIMAL, ImmutableList.of(), ImmutableList.of(), Optional.empty(), Optional.of(decimalType.precision()), Optional.of(decimalType.scale()), attributes));
-            case UUID:
-                attributes = ImmutableMap.<String, String>builder()
-                        .putAll(attributes)
-                        .put(ICEBERG_BINARY_TYPE, "UUID")
-                        .buildOrThrow();
-                return ImmutableList.of(new OrcType(OrcTypeKind.BINARY, ImmutableList.of(), ImmutableList.of(), Optional.empty(), Optional.empty(), Optional.empty(), attributes));
-            case STRUCT:
-                return toOrcStructType(nextFieldTypeIndex, (Types.StructType) type, attributes);
-            case LIST:
-                return toOrcListType(nextFieldTypeIndex, (Types.ListType) type, attributes);
-            case MAP:
-                return toOrcMapType(nextFieldTypeIndex, (Types.MapType) type, attributes);
-        }
-        throw new TrinoException(NOT_SUPPORTED, "Unsupported Iceberg type: " + type);
-    }
-
-    private static List<OrcType> toOrcStructType(int nextFieldTypeIndex, Types.StructType structType, Map<String, String> attributes)
-    {
-        nextFieldTypeIndex++;
-        List<OrcColumnId> fieldTypeIndexes = new ArrayList<>();
-        List<String> fieldNames = new ArrayList<>();
-        List<List<OrcType>> fieldTypesList = new ArrayList<>();
-        for (Types.NestedField field : structType.fields()) {
-            fieldTypeIndexes.add(new OrcColumnId(nextFieldTypeIndex));
-            fieldNames.add(field.name());
-            Map<String, String> fieldAttributes = ImmutableMap.<String, String>builder()
-                    .put(ORC_ICEBERG_ID_KEY, Integer.toString(field.fieldId()))
-                    .put(ORC_ICEBERG_REQUIRED_KEY, Boolean.toString(field.isRequired()))
-                    .buildOrThrow();
-            List<OrcType> fieldOrcTypes = toOrcType(nextFieldTypeIndex, field.type(), fieldAttributes);
-            fieldTypesList.add(fieldOrcTypes);
-            nextFieldTypeIndex += fieldOrcTypes.size();
-        }
-
-        ImmutableList.Builder<OrcType> orcTypes = ImmutableList.builder();
-        orcTypes.add(new OrcType(
-                OrcTypeKind.STRUCT,
-                fieldTypeIndexes,
-                fieldNames,
-                Optional.empty(),
-                Optional.empty(),
-                Optional.empty(),
-                attributes));
-        fieldTypesList.forEach(orcTypes::addAll);
-
-        return orcTypes.build();
-    }
-
-    private static List<OrcType> toOrcListType(int nextFieldTypeIndex, Types.ListType listType, Map<String, String> attributes)
-    {
-        nextFieldTypeIndex++;
-        Map<String, String> elementAttributes = ImmutableMap.<String, String>builder()
-                .put(ORC_ICEBERG_ID_KEY, Integer.toString(listType.elementId()))
-                .put(ORC_ICEBERG_REQUIRED_KEY, Boolean.toString(listType.isElementRequired()))
-                .buildOrThrow();
-        List<OrcType> itemTypes = toOrcType(nextFieldTypeIndex, listType.elementType(), elementAttributes);
-
-        List<OrcType> orcTypes = new ArrayList<>();
-        orcTypes.add(new OrcType(
-                OrcTypeKind.LIST,
-                ImmutableList.of(new OrcColumnId(nextFieldTypeIndex)),
-                ImmutableList.of("item"),
-                Optional.empty(),
-                Optional.empty(),
-                Optional.empty(),
-                attributes));
-
-        orcTypes.addAll(itemTypes);
-        return orcTypes;
-    }
-
-    private static List<OrcType> toOrcMapType(int nextFieldTypeIndex, Types.MapType mapType, Map<String, String> attributes)
-    {
-        nextFieldTypeIndex++;
-        Map<String, String> keyAttributes = ImmutableMap.<String, String>builder()
-                .put(ORC_ICEBERG_ID_KEY, Integer.toString(mapType.keyId()))
-                .put(ORC_ICEBERG_REQUIRED_KEY, Boolean.toString(true))
-                .buildOrThrow();
-        List<OrcType> keyTypes = toOrcType(nextFieldTypeIndex, mapType.keyType(), keyAttributes);
-        Map<String, String> valueAttributes = ImmutableMap.<String, String>builder()
-                .put(ORC_ICEBERG_ID_KEY, Integer.toString(mapType.valueId()))
-                .put(ORC_ICEBERG_REQUIRED_KEY, Boolean.toString(mapType.isValueRequired()))
-                .buildOrThrow();
-        List<OrcType> valueTypes = toOrcType(nextFieldTypeIndex + keyTypes.size(), mapType.valueType(), valueAttributes);
-
-        List<OrcType> orcTypes = new ArrayList<>();
-        orcTypes.add(new OrcType(
-                OrcTypeKind.MAP,
-                ImmutableList.of(new OrcColumnId(nextFieldTypeIndex), new OrcColumnId(nextFieldTypeIndex + keyTypes.size())),
-                ImmutableList.of("key", "value"),
-                Optional.empty(),
-                Optional.empty(),
-                Optional.empty(),
-                attributes));
-
-        orcTypes.addAll(keyTypes);
-        orcTypes.addAll(valueTypes);
-        return orcTypes;
     }
 }
